+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+- 🚀 [2017-05-12] `engine/components/numericLineGenerator`: update to v3.2.0.
+  - 🚀 New method: `beSad`. Updates the sad state for each number contained by the `Line`.
 - 🚀 [2017-05-12] `engine/containers/numberGenerator`: update to v3.2.0.
   - 🚀 Add new method `beSad`: it receives a `state` (`'start' | 'stop'`). Is used when the game has finish uncorrectly.
 - 🚀 [2017-05-12] `engine/helpers/customEases`: update to v1.1.0.